Updated the installation code and fixed typo
The repository no longer maintained linux-installer.sh but refers to it as install.sh. So changed the code on the readme file with the correct code. Additionally, there was a typo in the name of the document "requirements". I have fixed the name of the document to the correct one in another pull request, and have updated the code here in the readme file.