fix error for TimeZone
The previous syntax correctly expected the return of getTimeZone() and if
the client was linux you attribute a string with "Not Found" to the variable.

So it means the error was after the function call.

Following [1] you might just need to switch that value into a string.

Alternatively through the exceptions we can check for any return errors
of the getTimeZone() function. The value will then be a general "Not Found"

[1]: https://stackoverflow.com/a/62672362/19163681

Signed-off-by: Peppe289 <gsperanza204@gmail.com>