Fixed An Error Caused During bash install.sh command
ERROR: Invalid requirement: './requirements.txt'
Hint: It looks like a path. The path does exist. The argument you provided (./requirements.txt) appears to be a requirements file. If that is the case, use the '-r' flag to install the packages specified within it.
An error occurred! seems pip doesn't work.